晚上的commit

+++ LogisticManageSystem/src/main/java/edu/nju/lms/businessLogicService/FinanceReceiptblService.java
@@ -23,7 +23,7 @@ public interface FinanceReceiptblService {
 	 */
 	public ArrayList<ReceiptVO>  showReceiptVO (Calendar date, String department);
 	
-	public long getReceiptSum(Calendar date);
+	public double getReceiptSum(Calendar date);
 	
 	public ArrayList<ReceiptVO>  showReceiptList (Calendar date);
